Finish UserRegister handler
