+++ handlers/user/data.go
@@ -1,0 +1,6 @@
+package user
+
+import "github.com/gin-gonic/gin"
+
+func UserData(c *gin.Context) {
+}
