+++ handlers/user/login.go
@@ -1,0 +1,12 @@
+package user
+
+import "github.com/gin-gonic/gin"
+
+type ReqUserLogin struct {
+	Password string `json:"password" binding:"required"`
+	Email    string `json:"email" binding:"required"`
+}
+
+func UserLogin(c *gin.Context) {
+
+}
