+++ handlers/user/logout.go
@@ -1,0 +1,6 @@
+package user
+
+import "github.com/gin-gonic/gin"
+
+func UserLogout(c *gin.Context) {
+}
