+++ handlers/user/register.go
@@ -1,20 +1,12 @@
-package handlers
+package user
 
 import (
+	"STUOJ/db"
 	"STUOJ/model"
 	"github.com/gin-gonic/gin"
 	"log"
 	"net/http"
 )
-
-type ReqUserLogin struct {
-	Password string `json:"password" binding:"required"`
-	Email    string `json:"email" binding:"required"`
-}
-
-func UserLogin(c *gin.Context) {
-
-}
 
 type ReqUserRegister struct {
 	Username string `json:"username" binding:"required"`
@@ -25,26 +17,39 @@ type ReqUserRegister struct {
 func UserRegister(c *gin.Context) {
 	var req ReqUserRegister
 
+	// 参数绑定
 	err := c.ShouldBindBodyWithJSON(&req)
 	if err != nil {
 		log.Println(err)
 		c.JSON(http.StatusBadRequest, model.Response{
 			Code: 0,
-			Msg:  err.Error(),
+			Msg:  "参数错误",
 			Data: nil,
 		})
 		return
 	}
 
+	// 初始化用户
+	u := model.User{
+		Username: req.Username,
+		Password: req.Password,
+		Email:    req.Email,
+	}
+	err = db.SaveUser(u)
+	if err != nil {
+		log.Println(err)
+		c.JSON(http.StatusInternalServerError, model.Response{
+			Code: 0,
+			Msg:  "注册失败",
+			Data: nil,
+		})
+		return
+	}
+
+	// 返回结果
 	c.JSON(http.StatusOK, model.Response{
 		Code: 1,
-		Msg:  "OK",
-		Data: req,
+		Msg:  "注册成功",
+		Data: nil,
 	})
-}
-
-func UserLogout(c *gin.Context) {
-}
-
-func UserData(c *gin.Context) {
 }
